proxyprotocol use long long

+++ proxyprotocol.h
@@ -1,9 +1,7 @@
 #ifndef _PROXYPROTOCOL_H____
 #define _PROXYPROTOCOL_H____
 
-#include <stddef.h>
-
-extern int proxyprotocol_v1(char *, size_t);
-extern int proxyprotocol_v2(char *, size_t);
+extern long long proxyprotocol_v1(char *, long long);
+extern long long proxyprotocol_v2(char *, long long);
 
 #endif
